backend/requirements.txt
Flask
flask-cors
pymongo
python-dotenv
bcrypt
PyJWT
cloudinary
dnspython
gunicorn
